# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://dsm-1.tasklist.camunda.io/c7f8fc37-6771-4901-8c64-95ca7fd261e3
  description: Generated server url
...
paths:
...
  /v1/tasks/search:
    post:
      tags:
      - Task
      summary: Search tasks
      description: "Returns the list of tasks that satisfy search request params.<br><ul><li>If\
        \ an empty body is provided, all tasks are returned.</li><li>Only one of `[searchAfter,\
        \ searchAfterOrEqual, searchBefore, searchBeforeOrEqual]` search options must\
        \ be present in request.</li></ul>"
      operationId: searchTasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSearchRequest"
      responses:
        "200":
          description: On success returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskSearchResponse"
        "400":
          description: "An error is returned when more than one search parameters\
            \ among `[searchAfter, searchAfterOrEqual, searchBefore, searchBeforeOrEqual]`\
            \ are present in request"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/tasks/{taskId}:
    get:
      tags:
      - Task
      summary: Get a task
      description: Get one task by id. Returns task or error when task does not exist.
      operationId: getTaskById
      parameters:
      - name: taskId
        in: path
        description: The ID of the task.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: On success returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: An error is returned when the task with the `taskId` is not
            found.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User has no permission to access the task (Self-managed only).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          description: "An integer that represents the HTTP status code of the error\
            \ response. For example, 400 indicates a 'Bad Request' error, 404 indicates\
            \ a 'Not Found' error, and so on."
          format: int32
        message:
          type: string
          description: A string that provides a brief description of the error that
            occurred.
        instance:
          type: string
          description: "Error instance UUID for lookup (e.g., in log messages)."
...
    IncludeVariable:
      type: object
      properties:
        name:
          type: string
          description: The name of the variable.
        alwaysReturnFullValue:
          type: boolean
          description: Always return the full value of the variable?
          default: false
...
    DraftSearchVariableValue:
      type: object
      properties:
        value:
          type: string
          description: The value of the variable.
        isValueTruncated:
          type: boolean
          description: Does the `previewValue` contain the truncated value or full
            value?
        previewValue:
          type: string
          description: A preview of the variable's value. Limited in size.
      description: The draft value of the variable.
    VariableSearchResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the variable.
        name:
          type: string
          description: The name of the variable.
        value:
          type: string
          description: The value of the variable.
        isValueTruncated:
          type: boolean
          description: Does the `previewValue` contain the truncated value or full
            value?
        previewValue:
          type: string
          description: A preview of the variable's value. Limited in size.
        draft:
          $ref: "#/components/schemas/DraftSearchVariableValue"
    DateFilter:
      type: object
      properties:
        from:
          type: string
          description: Start date range to search from in date-time format outlined
            in section 5.6 of the RFC 3339 profile of the ISO 8601 standard.
          format: date-time
        to:
          type: string
          description: End date range to search to in date-time format outlined in
            section 5.6 of the RFC 3339 profile of the ISO 8601 standard.
          format: date-time
      description: A range of due dates for the tasks to search for.
    TaskByVariables:
      type: object
      properties:
        name:
          type: string
          description: The name of the variable.
        value:
          type: string
          description: "The value of the variable. When specifying the variable value,\
            \ it's crucial to maintain consistency with JSON values (serialization\
            \ for the complex objects such as list) and ensure that strings remain\
            \ appropriately formatted."
        operator:
          type: string
          description: "The comparison operator to use for the variable.<br>* `eq`:\
            \ Equals"
          enum:
          - eq
    TaskOrderBy:
      type: object
      properties:
        field:
          type: string
          enum:
          - completionTime
          - creationTime
          - followUpDate
          - dueDate
        order:
          type: string
          description: "* `ASC`: Ascending<br>* `DESC`: Descending"
          enum:
          - ASC
          - DESC
      description: Sort results by a specific field.
    TaskSearchRequest:
      type: object
      properties:
        state:
          type: string
          description: The state of the tasks.
          enum:
          - CREATED
          - COMPLETED
          - CANCELED
          - FAILED
        assigned:
          type: boolean
          description: Are the tasks assigned?
        assignee:
          type: string
          description: Who is assigned to the tasks?
        assignees:
          type: array
          description: The assignee is one of the given assignees.
          items:
            type: string
            description: The assignee is one of the given assignees.
        taskDefinitionId:
          type: string
          description: What's the BPMN flow node?
        candidateGroup:
          type: string
          description: Given group is in candidate groups list.
        candidateGroups:
          type: array
          description: At least one of the given groups is in candidate groups list.
          items:
            type: string
            description: At least one of the given groups is in candidate groups list.
        candidateUser:
          type: string
          description: Given user is in candidate user list.
        candidateUsers:
          type: array
          description: At least one of the given users is in candidate user list.
          items:
            type: string
            description: At least one of the given users is in candidate user list.
        processDefinitionKey:
          type: string
          description: Reference to process definition (renamed equivalent of TaskQuery.processDefinitionId
            field).
        processInstanceKey:
          type: string
          description: Reference to process instance (renamed equivalent of TaskQuery.processInstanceId
            field)
        pageSize:
          type: integer
          description: Size of tasks page (default = 50).
          format: int32
        followUpDate:
          $ref: "#/components/schemas/DateFilter"
        dueDate:
          $ref: "#/components/schemas/DateFilter"
        taskVariables:
          type: array
          description: "An array of filter clauses specifying the variables to filter\
            \ for.<br>If defined, the query returns only tasks to which all clauses\
            \ apply.<br>However, it's important to note that this filtering mechanism\
            \ is<br>designed to work exclusively with truncated variables. This means<br>variables\
            \ of a larger size are not compatible with this filter, and<br>attempts\
            \ to use them may result in inaccurate or incomplete query results."
          items:
            $ref: "#/components/schemas/TaskByVariables"
        tenantIds:
          type: array
          description: "An array of Tenant IDs to filter tasks. When multi-tenancy\
            \ is<br>enabled, tasks associated with the specified tenant IDs are returned;<br>if\
            \ disabled, this parameter is ignored."
          items:
            type: string
        sort:
          type: array
          description: An array of objects specifying the fields to sort the results
            by.
          items:
            $ref: "#/components/schemas/TaskOrderBy"
        searchAfter:
          type: array
          description: "Used to return a paginated result. Array of values that should\
            \ be copied from sortValues of one of the tasks from the current search\
            \ results page.<br>It enables the API to return a page of tasks that directly\
            \ follow the task identified by the provided values, with respect to the\
            \ sorting order."
          items:
            type: string
        searchAfterOrEqual:
          type: array
          description: "Used to return a paginated result. Array of values that should\
            \ be copied from sortValues of one of the tasks from the current search\
            \ results page.<br>It enables the API to return a page of tasks that directly\
            \ follow or are equal to the task identified by the provided values, with\
            \ respect to the sorting order."
          items:
            type: string
        searchBefore:
          type: array
          description: "Used to return a paginated result. Array of values that should\
            \ be copied from sortValues of one of the tasks from the current search\
            \ results page.<br>It enables the API to return a page of tasks that directly\
            \ precede the task identified by the provided values, with respect to\
            \ the sorting order."
          items:
            type: string
        searchBeforeOrEqual:
          type: array
          description: "Used to return a paginated result. Array of values that should\
            \ be copied from sortValues of one of the tasks from the current search\
            \ results page.<br>It enables the API to return a page of tasks that directly\
            \ precede or are equal to the task identified by the provided values,\
            \ with respect to the sorting order."
          items:
            type: string
        includeVariables:
          type: array
          description: An array used to specify a list of variable names that should
            be included in the response when querying tasks.<br>This field allows
            users to selectively retrieve specific variables associated with the tasks
            returned in the search results.
          items:
            $ref: "#/components/schemas/IncludeVariable"
        implementation:
          type: string
          enum:
          - JOB_WORKER
          - ZEEBE_USER_TASK
      description: Request object to search tasks by provided params.
    TaskSearchResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the task.
        name:
          type: string
          description: The name of the task.
        taskDefinitionId:
          type: string
          description: User Task ID from the BPMN definition.
        processName:
          type: string
          description: The name of the process.
        creationDate:
          type: string
          description: When was the task created (renamed equivalent of `Task.creationTime`
            field).
        completionDate:
          type: string
          description: When was the task completed (renamed equivalent of `Task.completionTime`
            field).
        assignee:
          type: string
          description: The username/id of who is assigned to the task.
        taskState:
          type: string
          description: The state of the task.
          enum:
          - CREATED
          - COMPLETED
          - CANCELED
          - FAILED
        sortValues:
          type: array
          description: Array of values to be copied into `TaskSearchRequest` to request
            for next or previous page of tasks.
          items:
            type: string
        isFirst:
          type: boolean
          description: A flag to show that the task is first in the current filter.
        formKey:
          type: string
          description: Reference to the task form.
        formId:
          type: string
          description: "Reference to the ID of a deployed form. If the form is not\
            \ deployed, this property is null."
        formVersion:
          type: integer
          description: "Reference to the version of a deployed form. If the form is\
            \ not deployed, this property is null."
          format: int64
        isFormEmbedded:
          type: boolean
          description: "Is the form embedded for this task? If there is no form, this\
            \ property is null."
        processDefinitionKey:
          type: string
          description: Reference to process definition (renamed equivalent of `Task.processDefinitionId`
            field).
        processInstanceKey:
          type: string
          description: Reference to process instance id (renamed equivalent of `Task.processInstanceId`
            field).
        tenantId:
          type: string
          description: The tenant ID associated with the task.
        dueDate:
          type: string
          description: The due date for the task.
          format: date-time
        followUpDate:
          type: string
          description: The follow-up date for the task.
          format: date-time
        candidateGroups:
          type: array
          description: The candidate groups for the task.
          items:
            type: string
        candidateUsers:
          type: array
          description: The candidate users for the task.
          items:
            type: string
        variables:
          type: array
          description: An array of the task's variables. Only variables specified
            in `TaskSearchRequest.includeVariables` are returned. Note that a variable's
            draft value is not returned in `TaskSearchResponse`.
          items:
            $ref: "#/components/schemas/VariableSearchResponse"
        context:
          type: string
          description: The context variable (from modeler) of the task.
        implementation:
          type: string
          enum:
          - JOB_WORKER
          - ZEEBE_USER_TASK
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the task.
        name:
          type: string
          description: The name of the task.
        taskDefinitionId:
          type: string
          description: User Task ID from the BPMN definition.
        processName:
          type: string
          description: The name of the process.
        creationDate:
          type: string
          description: When was the task created (renamed equivalent of `Task.creationTime`
            field).
        completionDate:
          type: string
          description: When was the task completed (renamed equivalent of `Task.completionTime`
            field).
        assignee:
          type: string
          description: The username/id of who is assigned to the task.
        taskState:
          type: string
          description: The state of the task.
          readOnly: true
          enum:
          - CREATED
          - COMPLETED
          - CANCELED
          - FAILED
        formKey:
          type: string
          description: Reference to the task form.
        formId:
          type: string
          description: "Reference to the ID of a deployed form. If the form is not\
            \ deployed, this property is null."
        formVersion:
          type: integer
          description: "Reference to the version of a deployed form. If the form is\
            \ not deployed, this property is null."
          format: int64
        isFormEmbedded:
          type: boolean
          description: "Is the form embedded for this task? If there is no form, this\
            \ property is null."
        processDefinitionKey:
          type: string
          description: Reference to process definition (renamed equivalent of `Task.processDefinitionId`
            field).
        processInstanceKey:
          type: string
          description: Reference to process instance id (renamed equivalent of `Task.processInstanceId`
            field).
        tenantId:
          type: string
          description: The tenant ID associated with the task.
        dueDate:
          type: string
          description: The due date for the task.
          format: date-time
        followUpDate:
          type: string
          description: The follow-up date for the task.
          format: date-time
        candidateGroups:
          type: array
          description: The candidate groups for the task.
          items:
            type: string
        candidateUsers:
          type: array
          description: The candidate users for the task.
          items:
            type: string
        implementation:
          type: string
          enum:
          - JOB_WORKER
          - ZEEBE_USER_TASK
...